+++ src/components/pause.rs
@@ -5,6 +5,7 @@ use serenity::all::{ComponentInteraction, Context};
 use tracing::{event, Level};
 
 use crate::i18n::t;
+use crate::utils::delete_player_message;
 use crate::{utils, PLAYER_MANAGER};
 
 /// Executes the `pause` command.
@@ -28,6 +29,7 @@ pub async fn execute<'a>(context: &Context, interaction: &ComponentInteraction) 
 
     let player_state = manager
         .get_voice_channel_id(guild_id)
+        .await
         .zip(manager.get_pause(guild_id));
 
     if let Some((my_channel_id, paused)) = player_state {
@@ -50,6 +52,8 @@ pub async fn execute<'a>(context: &Context, interaction: &ComponentInteraction) 
             Cow::borrowed(t(&interaction.locale, "error.not_in_voice_channel"))
         }
     } else {
+        delete_player_message(context, interaction).await;
+
         Cow::borrowed(t(&interaction.locale, "error.player_not_exists"))
     }
 }